+++ R/geom_tad.R
@@ -2,11 +2,12 @@
 #'
 #' @param matrix Matrix (n x n) contains contact map information.
 #' @param granges The rownames and colnames information of matrix.
-#' @param zlim The maximum value of color to plot. Larger values will be truncated to this value.
-#' @param color.bias Bias parameter for color palette, it is the same as \code{bias} parameter of \code{\link{colorRampPalette}}. Default: 1.
-#' @param color.ramp A name of a color palette, choose from rownames of \code{\link{brewer.pal.info}}. Default: YlOrRd.
-#' @param color.palette A vector of colors. Overwrites the palette for \code{color.ramp}. Default: NULL.
-#' @param smooth.func An optional smoothing function for the matrix. Default: NULL.
+#' @param color.palette One of the RColorbrewer or viridis colour palettes.
+#' Parameter of \code{\link{Brick_vizart_plot_heatmap}}. Default: NULL.
+#' @param value.cut If present, values beyond a certain quantile will be capped to that quantile.
+#' Parameter of \code{\link{Brick_vizart_plot_heatmap}}. Default: NULL.
+#' @param transform.fun If any sort of transformations should be applied to the data before plotting.
+#' Parameter of \code{\link{Brick_vizart_plot_heatmap}}. Default: NULL.
 #' @param plot.space Top and bottom margin. Default: 0.1.
 #' @param plot.height The relative height of contact map to coverage plot. Default: 1.
 #' @param top Logical value, whether to place this plot on the coverage plot. Default: TRUE.
@@ -15,37 +16,20 @@
 #' @return Plot.
 #' @importFrom magrittr %>%
 #' @importFrom GenomicRanges GRanges
-#' @importFrom IRanges IRanges
-#' @importFrom GenomeMatrix matrixPlotter
+#' @importFrom IRanges IRanges findOverlaps subsetByOverlaps
+#' @import HiCBricks
 #' @importFrom ggplot2 ggplot_add ggplot labs theme_classic theme element_blank element_rect
-#' element_text margin scale_y_continuous scale_x_continuous coord_cartesian
+#' element_text margin scale_y_continuous scale_x_continuous
 #' @importFrom patchwork wrap_plots
 #'
 #' @export
 #'
-#' @examples
-#' library(ggcoverage)
-#' library(HiCDataHumanIMR90)
-#' data(Dixon2012_IMR90, package = "HiCDataHumanIMR90")
-#' mat <- as.matrix(hic_imr90_40@.Data[[1]]@intdata)
-#' granges <- hic_imr90_40@.Data[[1]]@xgi
-#' # prepare coverage dataframe
-#' df <- data.frame(
-#'   seqnames = "chr1", start = seq(from = 50000000, to = 59999000, by = 1000),
-#'   end = seq(from = 50001000, to = 60000000, by = 1000), score = sample(1:100, 10000, replace = TRUE),
-#'   Type = "Example", Group = "Example"
-#' )
-#' # create plot
-#' ggcoverage(
-#'   data = df, color = "grey", region = "chr1:50000000-56000000",
-#'   mark.region = NULL, range.position = "out"
-#' ) +
-#'   geom_tad(matrix = log2(mat + 1), granges = granges, zlim = 5, color.palette = c("blue", "red"))
-geom_tad <- function(matrix, granges, zlim = NULL, color.bias = 1, color.ramp = "YlOrRd",
-                     color.palette = NULL, smooth.func = NULL, plot.space = 0.1, plot.height = 1, top = TRUE, show.rect = FALSE) {
+geom_tad <- function(matrix, granges, color.palette = NULL, value.cut = NULL,
+                     transform.fun = NULL, plot.space = 0.1, plot.height = 1,
+                     top = TRUE, show.rect = FALSE) {
   structure(list(
-    matrix = matrix, granges = granges, zlim = zlim, color.bias = color.bias,
-    color.ramp = color.ramp, color.palette = color.palette, smooth.func = smooth.func,
+    matrix = matrix, granges = granges, color.palette = color.palette,
+    value.cut = value.cut, transform.fun = transform.fun,
     plot.space = plot.space, plot.height = plot.height, top = top, show.rect = show.rect
   ),
   class = "tad"
@@ -76,26 +60,66 @@ ggplot_add.tad <- function(object, plot, object_name) {
   # get parameters
   matrix <- object$matrix
   granges <- object$granges
-  zlim <- object$zlim
-  color.bias <- object$color.bias
-  color.ramp <- object$color.ramp
   color.palette <- object$color.palette
-  smooth.func <- object$smooth.func
+  value.cut <- object$value.cut
+  transform.fun <- object$transform.fun
   plot.space <- object$plot.space
   plot.height <- object$plot.height
   top <- object$top
   show.rect <- object$show.rect
-  # plot
-  tad.plot <- GenomeMatrix::matrixPlotter(
-    mat = matrix, granges = granges, plotGR = plot.range.gr,
-    extend = 0, heightProp = 1 / 2, zlim = zlim, colorBias = color.bias,
-    highlight = NULL, colRamp = color.ramp, colPalette = color.palette,
-    smoothFilt = smooth.func
+
+  # create tmp directory
+  out.dir <- file.path(tempdir(), "ggcoverage")
+  if (!dir.exists(out.dir)) {
+    dir.create(out.dir)
+  }
+
+  # filter used matrix and bin
+  used.range.index <- as.data.frame(IRanges::findOverlaps(granges, plot.range.gr))[, "queryHits"]
+  matrix.used <- matrix[used.range.index, used.range.index]
+  used.range.df <- as.data.frame(IRanges::subsetByOverlaps(x = granges, ranges = plot.range.gr))
+
+  # write used mat and bin to file
+  write.table(
+    x = used.range.df[c(1, 2, 3)], file = file.path(out.dir, "used.range.txt"),
+    sep = "\t", quote = F, row.names = F, col.names = F
   )
+  write.table(
+    x = as.matrix(matrix.used), file = file.path(out.dir, "used.matrix.txt"),
+    sep = "\t", quote = F, row.names = F, col.names = F
+  )
+
+  # create brick container
+  BrickContainer <- HiCBricks::Create_many_Bricks(
+    BinTable = file.path(out.dir, "used.range.txt"),
+    bin_delim = "\t", output_directory = out.dir, file_prefix = "ggcoverage",
+    experiment_name = "ggcoverage", resolution = 100000,
+    remove_existing = TRUE, impose_discontinuity = FALSE
+  )
+
+  # load matrix
+  Load.matrix <- HiCBricks::Brick_load_matrix(
+    Brick = BrickContainer, chr1 = as.character(plot.range.chr),
+    chr2 = as.character(plot.range.chr),
+    matrix_file = file.path(out.dir, "used.matrix.txt"), delim = "\t",
+    remove_prior = TRUE, resolution = 100000
+  )
+
+  # create plot
+  tad.plot <-
+    HiCBricks::Brick_vizart_plot_heatmap(
+      File = file.path(out.dir, "ggcoverage_hic_plot.pdf"),
+      Bricks = list(BrickContainer),
+      x_coords = paste(plot.range.chr, plot.range.start, plot.range.end, sep = ":"),
+      y_coords = paste(plot.range.chr, plot.range.start, plot.range.end, sep = ":"),
+      resolution = 100000, FUN = transform.fun, value_cap = value.cut,
+      distance = NULL, legend_title = "", palette = color.palette,
+      width = 10, height = 10, rotate = TRUE, return_object = TRUE
+    )
   # add theme
   tad.final.plot <- tad.plot +
     labs(y = "ContactMap") +
-    theme_tad(x.range = c(plot.range.start, plot.range.end), margin.len = plot.space, show.rect = show.rect)
+    theme_tad(margin.len = plot.space, show.rect = show.rect)
   # assemble plot
   if (top) {
     patchwork::wrap_plots(tad.final.plot + theme(plot.margin = margin(t = plot.space, b = plot.space)),