+++ R/theme_ggcoverage.R
@@ -290,7 +290,7 @@ theme_peak <- function(margin.len, x.range) {
   )
 }
 
-#' Theme for geom_tad.
+#' Theme for geom_tad2.
 #'
 #' @param x.range X axis ranges.
 #' @param margin.len Top and bottom margin.
@@ -299,7 +299,7 @@ theme_peak <- function(margin.len, x.range) {
 #' @return List of layers.
 #' @importFrom ggplot2 theme_classic theme element_blank element_text element_rect margin
 #' scale_x_continuous scale_y_continuous coord_cartesian
-theme_tad <- function(x.range, margin.len, show.rect) {
+theme_tad2 <- function(x.range, margin.len, show.rect) {
   if (show.rect) {
     list(
       theme_classic(),
@@ -335,9 +335,60 @@ theme_tad <- function(x.range, margin.len, show.rect) {
         legend.title = element_blank(),
         plot.margin = margin(t = margin.len, b = margin.len)
       ),
-      scale_y_continuous(expand = c(0, 0)),
+      scale_y_continuous(expand = c(0, 0), position = "right"),
       scale_x_continuous(expand = c(0, 0)),
       coord_cartesian(xlim = x.range)
+    )
+  }
+}
+
+#' Theme for geom_tad.
+#'
+#' @param margin.len Top and bottom margin.
+#' @param show.rect Logical value, whether to add rect border to the plot.
+#'
+#' @return List of layers.
+#' @importFrom ggplot2 theme_classic theme element_blank element_text element_rect margin
+#' scale_x_continuous scale_y_continuous
+theme_tad <- function(margin.len, show.rect) {
+  if (show.rect) {
+    list(
+      theme_classic(),
+      theme(
+        axis.line.x = element_blank(),
+        axis.line.y = element_blank(),
+        axis.text.y = element_blank(),
+        axis.title.y.right = element_text(color = "black", angle = 90, vjust = 0.5),
+        axis.ticks.y = element_blank(),
+        axis.text.x = element_blank(),
+        axis.title.x = element_blank(),
+        axis.ticks.x = element_blank(),
+        legend.title = element_blank(),
+        plot.title = element_blank(),
+        panel.border = element_rect(colour = "black", fill = NA, size = 1),
+        plot.margin = margin(t = margin.len, b = margin.len)
+      ),
+      scale_y_continuous(expand = c(0, 0), position = "right"),
+      scale_x_continuous(expand = c(0, 0))
+    )
+  } else {
+    list(
+      theme_classic(),
+      theme(
+        axis.line.x = element_blank(),
+        axis.line.y = element_blank(),
+        axis.text.y = element_blank(),
+        axis.title.y.right = element_text(color = "black", angle = 90, vjust = 0.5),
+        axis.ticks.y = element_blank(),
+        axis.text.x = element_blank(),
+        axis.title.x = element_blank(),
+        axis.ticks.x = element_blank(),
+        legend.title = element_blank(),
+        plot.title = element_blank(),
+        plot.margin = margin(t = margin.len, b = margin.len)
+      ),
+      scale_y_continuous(expand = c(0, 0), position = "right"),
+      scale_x_continuous(expand = c(0, 0))
     )
   }
 }