+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggcoverage 1.0.0
 ## Major changes
-* Supporting Hi-C visualization (`geom_tad`).
+* Supporting Hi-C visualization (`geom_tad` and `geom_tad2`).
 * Supporting genome region interaction visualization (`geom_link`).
 * Supporting CNV visualization (`geom_cnv`).